통계 자료형 변경 (#22)

+++ sberp/src/main/java/egovframework/sts/dto/StsDTO.java
@@ -16,7 +16,7 @@ public class StsDTO {
   @ApiModelProperty(value="계산서발행 금액 ", example = "")
   private int billAmt;
   @ApiModelProperty(value="납부완료 금액 ", example = "")
-  private int payAmt;
+  private long payAmt;
   
   @ApiModelProperty(value="시험부서별 데이터 ", example = "")
   private List<TestTypeList> testTypeList;
@@ -34,36 +34,36 @@ public class StsDTO {
     @ApiModelProperty(value="시험접수 건수 ", example = "")
     private int inCnt;
     @ApiModelProperty(value="시험접수 금액 ", example = "")
-    private int inAmt;  
+    private long inAmt;  
     
     @ApiModelProperty(value="시험중 건수 ", example = "")
     private int ingCnt;
     @ApiModelProperty(value="시험중 금액 ", example = "")
-    private int ingAmt;
+    private long ingAmt;
     @ApiModelProperty(value="디버깅 건수 ", example = "")
     private int debCnt;
     @ApiModelProperty(value="디버깅 금액 ", example = "")
-    private int debAmt;
+    private long debAmt;
     @ApiModelProperty(value="홀딩 건수 ", example = "")
     private int holCnt;
     @ApiModelProperty(value="홀딩 금액 ", example = "")
-    private int holAmt;
+    private long holAmt;
     @ApiModelProperty(value="시험완료 건수 ", example = "")
     private int endCnt;
     @ApiModelProperty(value="시험완료 금액 ", example = "")
-    private int endAmt;
+    private long endAmt;
     @ApiModelProperty(value="성적서발급완료 건수 ", example = "")
     private int repCnt;
     @ApiModelProperty(value="성적서발급완료 금액 ", example = "")
-    private int repAmt;
+    private long repAmt;
     @ApiModelProperty(value=" ", example = "")
     private int etcCnt;
     @ApiModelProperty(value=" ", example = "")
-    private int etcAmt;
+    private long etcAmt;
     @ApiModelProperty(value=" ", example = "")
     private int cnt;
     @ApiModelProperty(value=" ", example = "")
-    private int amt;
+    private long amt;
     
   }
 }
